(GH-1732) Unify Nuget content resolvers.
- Use NuGet.Client on both full framework and .NET Core
- Fixes #1732
