+++ src/Cake.NuGet/Cake.NuGet.csproj
@@ -21,14 +21,13 @@
     <ProjectReference Include="..\Cake.Core\Cake.Core.csproj" />
   </ItemGroup>
 
-  <!-- .NET Core packages -->
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.6' ">
+  <!-- Global packages -->
+  <ItemGroup>
     <PackageReference Include="NuGet.Client" Version="4.0.0" />
   </ItemGroup>
 
   <!-- .NET Framework packages -->
   <ItemGroup Condition=" '$(TargetFramework)' == 'net46' ">
-    <PackageReference Include="NuGet.Core" Version="2.14.0" />
     <Reference Include="System" />
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
